refactor(dom): extract textOffsets helpers from DomModel
